pollingRoomCondition-lambda-sam/pollingRoomCondition/requirements.txt
requests
boto3
datetime
aws-psycopg2
